pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "racing-mcp-server"
version = "1.0.0"
description = "MCP server for The Racing API — connects Claude to live horse racing data"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0,<2.0.0",
    "httpx>=0.27.0,<1.0.0",
    "starlette>=0.41.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "cachetools>=5.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
]

[project.scripts]
racing-mcp = "racing_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/racing_mcp"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
